@@ -14,20 +14,20 @@
   },
   "devDependencies": {
     "@types/bootstrap": "^3.3.32",
-    "@types/chai": "^3.4.34",
-    "@types/express": "^4.0.34",
-    "@types/jquery": "^2.0.34",
-    "@types/mathjs": "0.0.31",
-    "@types/redis": "^0.12.34",
+    "@types/chai": "^3.4.35",
+    "@types/express": "^4.0.35",
+    "@types/jquery": "^2.0.40",
+    "@types/mathjs": "0.0.34",
+    "@types/redis": "^0.12.36",
     "@types/serve-index": "^1.7.28",
-    "@types/socket.io": "^1.4.27",
+    "@types/socket.io": "^1.4.28",
     "@types/socket.io-client": "^1.4.29",
     "@types/socket.io-redis": "^1.0.21",
-    "@types/systemjs": "^0.19.33",
+    "@types/systemjs": "^0.20.1",
     "@types/uuid": "^2.0.29",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-es2015": "^6.22.0",
     "bison": "^1.1.1",
-    "browser-sync": "^2.18.5",
+    "browser-sync": "^2.18.8",
     "core-js": "^2.4.1",
     "del": "^2.2.2",
     "eslint-plugin-json": "^1.2.0",
@@ -38,17 +38,17 @@
     "gulp-eslint": "^3.0.1",
     "gulp-if": "^2.0.2",
     "gulp-ignore": "^2.0.2",
-    "gulp-sass": "^3.0.0",
-    "gulp-sourcemaps": "^2.2.0",
-    "gulp-tslint": "^7.0.1",
-    "gulp-typescript": "^3.1.3",
-    "gulp-uglify": "^2.0.0",
+    "gulp-sass": "^3.1.0",
+    "gulp-sourcemaps": "^2.4.1",
+    "gulp-tslint": "^7.1.0",
+    "gulp-typescript": "^3.1.5",
+    "gulp-uglify": "^2.0.1",
     "minimist": "^1.2.0",
     "nodemon": "^1.11.0",
-    "systemjs": "^0.19.41",
+    "systemjs": "^0.20.9",
     "tether": "^1.4.0",
-    "tslint": "^4.2.0",
-    "typescript": "^2.2.0-dev.20161228"
+    "tslint": "^4.5.1",
+    "typescript": "^2.3.0-dev.20170302"
   },
   "scripts": {
     "clean": "gulp clean",
@@ -58,39 +58,38 @@
     "server": "gulp server"
   },
   "dependencies": {
-    "@angular/common": "^2.4.1",
-    "@angular/compiler": "^2.4.1",
-    "@angular/core": "^2.4.1",
-    "@angular/forms": "^2.4.1",
-    "@angular/http": "^2.4.1",
-    "@angular/platform-browser": "^2.4.1",
-    "@angular/platform-browser-dynamic": "^2.4.1",
-    "@angular/router": "^3.4.1",
-    "@angular/upgrade": "^2.4.1",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.15",
+    "@angular/common": "^2.4.9",
+    "@angular/compiler": "^2.4.9",
+    "@angular/core": "^2.4.9",
+    "@angular/forms": "^2.4.9",
+    "@angular/http": "^2.4.9",
+    "@angular/platform-browser": "^2.4.9",
+    "@angular/platform-browser-dynamic": "^2.4.9",
+    "@angular/router": "^3.4.9",
+    "@angular/upgrade": "^2.4.9",
     "angular2-auto-scroll": "^1.0.12",
-    "body-parser": "^1.15.2",
+    "body-parser": "^1.17.0",
     "bootstrap": "^4.0.0-alpha.4",
     "compression": "^1.6.2",
     "core-js": "^2.4.1",
-    "express": "^4.14.0",
-    "express-ws": "^2.0.0",
+    "express": "^4.15.0",
+    "express-ws": "^3.0.0",
     "jquery": "^3.1.1",
-    "mathjs": "^3.8.1",
+    "mathjs": "^3.9.3",
     "moment": "^2.17.1",
-    "redis": "^2.6.3",
-    "reflect-metadata": "^0.1.9",
-    "regenerator-runtime": "^0.10.1",
-    "rxjs": "^5.0.2",
+    "redis": "^2.6.5",
+    "reflect-metadata": "^0.1.10",
+    "regenerator-runtime": "^0.10.3",
+    "rxjs": "^5.2.0",
     "serve-index": "^1.8.0",
-    "socket.io-redis": "^2.0.1",
-    "source-map-support": "^0.4.8",
+    "socket.io-redis": "^4.0.0",
+    "source-map-support": "^0.4.11",
     "sqlite3": "^3.1.8",
-    "systemjs": "^0.19.41",
+    "systemjs": "^0.20.9",
     "tether": "^1.4.0",
-    "tslib": "^1.4.0",
-    "typeorm": "0.0.5",
+    "tslib": "^1.6.0",
+    "typeorm": "0.0.9",
     "uuid": "^3.0.1",
-    "zone.js": "^0.7.4"
+    "zone.js": "^0.7.7"
   }
 }
