@@ -1,12 +1,8 @@
-import Square from './square';
-
-declare global {
-	export interface BoardMouseEvent {
-		square: Square|null;
-		position: {
-			screen: Point,
-			canvas: Point,
-			board: Point|null
-		}
+declare interface BoardMouseEvent {
+	square: Square|null;
+	position: {
+		screen: Point,
+		canvas: Point,
+		board: Point|null
 	}
 }
