@@ -1,6 +1,6 @@
-import MapRepository from './map';
+import { MapRepository } from './map';
 
-export default class SessionRepository extends MapRepository<SessionRecord> {
+export class SessionRepository extends MapRepository<SessionRecord> {
 	public constructor() {
 		super( 'sessionId' );
 	}
