@@ -1,6 +1,6 @@
-import MapRepository from './map';
+import { MapRepository } from './map';
 
-export default class UserRepository extends MapRepository<UserRecord> {
+export class UserRepository extends MapRepository<UserRecord> {
 	public constructor() {
 		super( 'userId' );
 	}
