@@ -1,4 +1,4 @@
-export default class Bounds {
+export class Bounds {
 	public constructor( public readonly left: number, public readonly top: number, public readonly width: number, public readonly height: number ) {}
 
 	public get bottom() {
